cleanup: bot.py - Reorganising imports, extracting magic variables and removing unused imports

+++ src/bot.py
@@ -1,19 +1,15 @@
-from dataclasses import dataclass
-import discord
-from discord.ext import commands
-import sqlite3 as sql
-import time
-from utils import *
-from constants import *
-import random
-import json
-import re
-import threading
-import asyncio
 import os
-from dotenv import load_dotenv
+import random
+import re
+import time
 from datetime import datetime, timedelta, timezone
 
+import discord
+from discord.ext import commands
+from dotenv import load_dotenv
+
+from constants import *
+from utils import *
 
 load_dotenv()
 token = os.getenv("bot-token")
@@ -46,7 +42,7 @@ class InfoSuggestModal(discord.ui.Modal, title="Suggestion Form"):
     )
 
     async def on_submit(self, interaction: discord.Interaction):
-        with open("src/info_suggestions.txt", "a") as f:
+        with open("src/info_suggestions.txt", "a") as f:  # noqa: ASYNC230 -- If the json grows or this bot scales to more servers you should move to an async system
             f.write(f"{self.command.value} | {self.output.value}\n")
             await interaction.response.send_message("Suggestion sent!", ephemeral=True)
 
@@ -62,20 +58,21 @@ async def suggest(ctx):
 
 @bot.command()
 async def info(ctx, *, topic):
+    data = get_json("general_info.json")["info_command"]
     if topic == "help":
         message = "Available info commands: "
-        for info in get_json("general_info.json")["info_command"].keys():
-            message += f"`{info}`, "
+        for info_field in data:
+            message += f"`{info_field}`, "
 
         message = await ctx.channel.send(message[:-2], suppress=True)
         return
 
     try:
-        response = get_json("general_info.json")["info_command"][topic]
+        response = data[topic]
 
         if not isinstance(response, str):
             response = response[random.randint(0, len(response)-1)]
-    except:
+    except (KeyError, ValueError):
         response = f"Found no info on {topic}. All info commands can be found with `>info help`\nYou can also suggest a new one using `>suggest`"
 
     await ctx.channel.send(response)
@@ -87,7 +84,7 @@ async def chatrevive(ctx):
     global last_ping
 
     if (last_ping + CHAT_REVIVE_TIMEOUT < time.time()):
-        role = ctx.guild.get_role(1529912646248435803)
+        role = ctx.guild.get_role(1529912646248435803) # FIX: magic variable
         await ctx.channel.send(f"{role.mention}")
         last_ping = time.time()
     else:
@@ -105,15 +102,20 @@ async def files(ctx):
 @bot.listen()
 async def on_message(message):
     # If user is mod dont react to the message in hall of fame submissions
-    if user_has_role(message.author, MOD_ROLE_ID):
-        if message.content.startswith("!NR"):
-            return
+    if (
+        user_has_role(message.author, MOD_ROLE_ID)
+        and message.content.startswith("!NR")
+    ):
+        return
 
-    if message.channel.id == HOF_SUBMISSION_CHANNEL_ID:
-        if len(message.attachments) == 0:
-            if not user_has_role(message.author, MOD_ROLE_ID):
-                await message.delete()
-                return
+    if (
+        message.channel.id == HOF_SUBMISSION_CHANNEL_ID
+        and len(message.attachments) == 0
+        and not user_has_role(message.author, MOD_ROLE_ID)
+    ):
+        await message.delete()
+        return
+
 
     if message.channel.id == HOF_SUBMISSION_CHANNEL_ID:
         await message.add_reaction("\u2b50")
@@ -125,12 +127,12 @@ async def on_raw_reaction_add(payload):
         return
     if payload.channel_id != HOF_SUBMISSION_CHANNEL_ID:
         return
-    if str(payload.emoji) != "\u2b50":
+    if str(payload.emoji) != "\u2b50": # Star Emoji
         return
     channel = bot.get_channel(payload.channel_id)
     message = await channel.fetch_message(payload.message_id)
     reaction = discord.utils.get(message.reactions, emoji="\u2b50")
-    if reaction and reaction.count <= HOF_SUBMISSION_REACTIONS:
+    if not reaction or reaction.count <= HOF_SUBMISSION_REACTIONS:
         return
     if db.hof_message_exists(reaction.message.id):
         return
@@ -165,7 +167,7 @@ async def on_member_join(member):
     if role is not None:
         await member.add_roles(role, reason="Auto-role on join")
 
-@bot.listen
+@bot.listen()
 async def on_message_edit(before, after):
     keywords = ["*you're", "you're*", "*your", "your*"]
     for keyword in keywords:
@@ -201,7 +203,7 @@ async def on_message(message):
             print(2)
         return
 
-    date = message.created_at.timestamp()
+    # date = message.created_at.timestamp() # Unused var (not sure if you plan to use it for smt later)
 
     if message.author.id != ARCANE_USER_ID:
         return
